Moving all non page items to a new sub folder "support" for the ui elements
the ui folder was getting overly full
