+++ src/main/java/sharkbread/ui/PageBrowseRecipe.java
@@ -8,6 +8,7 @@
 package sharkbread.ui;
 
 import sharkbread.database.Recipe;
+import sharkbread.ui.support.PopUp;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
