+++ src/main/java/sharkbread/ui/PageDisplayRecipe.java
@@ -11,6 +11,8 @@ package sharkbread.ui;
 
 import sharkbread.database.Ingredient;
 import sharkbread.database.Recipe;
+import sharkbread.ui.support.ImportExport;
+import sharkbread.ui.support.PopUp;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
