+++ src/main/java/sharkbread/ui/support/EntryList.java
@@ -1,4 +1,4 @@
-package sharkbread.ui;
+package sharkbread.ui.support;
 
 import java.awt.GridLayout;
 import java.util.ArrayList;
